It's short

+++ CustomEnchant/src/fr/enchantments/custom/helper/MathHelper.java
@@ -8,7 +8,7 @@ public abstract class MathHelper {
 	 * @param i the integer to randomize
 	 * @return an integer between 1 and <code>i</code>
 	 */
-	public static int randomize(int i) {
-		return (int)(Math.random()*i+1);
+	public static short randomize(short i) {
+		return (short)(Math.random()*i+1);
 	}
 }
